CRUD ressource and ressource type

+++ src/resource/resource.service.ts
@@ -1,29 +1,103 @@
-import { Injectable } from "@nestjs/common";
+import { ForbiddenException, Injectable } from "@nestjs/common";
 // import { CreateResourceDto } from './dto/create-resource.dto';
 import { PrismaService } from "src/prisma/prisma.service";
 import { CreateResourceDto, UpdateResourceDto } from "./dto";
 import handleDeleteOnRestrictResponse from "src/utils/handleDeleteOnRestrictResponse";
+import { Response } from "express";
 
 @Injectable()
 export class ResourceService {
   constructor(private prisma: PrismaService) {}
-  create(dto: CreateResourceDto) {
-    return this.prisma.ressource.create({ data: { ...dto } });
+
+  async create(dto: CreateResourceDto, res: Response) {
+    try {
+      // Tester si l'id de la catégorie assignée existe, sinon throw une erreur
+
+      const data = await this.prisma.ressource.create({ data: { ...dto } });
+
+      return res.status(res.statusCode).json({
+        status: res.statusCode,
+        success: true,
+        message: "La ressource a été correctement créée",
+        // data: data,
+      });
+    } catch (error) {
+      console.log("ERROR: " + error.message);
+      return res.status(error.status).json({
+        status: error.status,
+        success: false,
+        message: error.message,
+        // error: { error: "Database connection error" },
+      });
+    }
   }
 
-  async findAll() {
-    return this.prisma.ressource.findMany();
+  async findAll(res: Response) {
+    try {
+      const data = await this.prisma.ressource.findMany();
+      return res.status(res.statusCode).json({
+        status: res.statusCode,
+        success: true,
+        message: "La liste a été récupéré",
+        data: data,
+      });
+    } catch (error) {
+      console.log("ERROR: " + error.message);
+      return res.status(error.status).json({
+        status: error.status,
+        success: false,
+        message: error.message,
+        // error: { error: "Database connection error" },
+      });
+    }
   }
 
-  findOne(id: string) {
-    return this.prisma.ressource.findUnique({ where: { id: id } });
+  async findOne(id: string, res: Response) {
+    try {
+      const data = await this.prisma.ressource.findUnique({
+        where: { id: id },
+      });
+
+      if (!data) {
+        throw new ForbiddenException("La ressurce n'a pas été trouvé.");
+      }
+      return res.status(res.statusCode).json({
+        status: res.statusCode,
+        success: true,
+        message: "La liste a été récupéré",
+        data: data,
+      });
+    } catch (error) {
+      console.log("ERROR: " + error.message);
+      return res.status(error.status).json({
+        status: error.status,
+        success: false,
+        message: error.message,
+        // error: { error: "Database connection error" },
+      });
+    }
   }
 
   async update(id: string, dto: UpdateResourceDto) {
-    const typeOfressource = await this.prisma.ressource.findFirst({
+    const listOfRessourcesTypes = await this.prisma.ressource_type.findMany();
+
+    const typeOfressource = await this.prisma.ressource.findUnique({
       where: { id: id },
       include: { Ressource_type: true },
     });
+
+    for (const ressourceType of listOfRessourcesTypes) {
+      if (typeOfressource.Ressource_type.id === ressourceType.id) {
+        await this.prisma.ressource.update({
+          where: { id: id },
+          data: {
+            name: dto.name,
+            price: dto.price,
+            acquisition_date: dto.acquisition_date,
+          },
+        });
+      }
+    }
 
     if (typeOfressource.Ressource_type.name === "material") {
       return this.prisma.ressource.update({
@@ -48,14 +122,40 @@ export class ResourceService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: string, res: Response) {
     try {
-      return await this.prisma.ressource.delete({ where: { id: id } });
+      const isRessourceExist = await this.prisma.ressource.findUnique({
+        where: { id: id },
+      });
+
+      if (!isRessourceExist) {
+        throw new ForbiddenException("La ressource n'a pas été trouvé.");
+      }
+
+      const data = await this.prisma.ressource.delete({ where: { id: id } });
+      if (!data) {
+        throw new ForbiddenException("La ressource n'a pas pu être supprimé.");
+      }
+      return res.status(res.statusCode).json({
+        status: res.statusCode,
+        success: true,
+        message: "La liste a été récupéré",
+        data: data,
+      });
     } catch (error) {
-      return handleDeleteOnRestrictResponse(this.prisma, id, "ressource", [
-        "comment",
-        "trucquidevraitpasmarcher",
-      ]);
+      console.log("ERROR: " + error.message);
+      const errorData = await handleDeleteOnRestrictResponse(
+        this.prisma,
+        id,
+        "ressource",
+        ["comment", "trucquidevraitpasmarcher"],
+      );
+      return res.status(error.status).json({
+        status: error.status,
+        success: false,
+        message: error.message,
+        error: errorData,
+      });
     }
   }
 }
